Allow packets to travel in both directions.
This required creating a new Direction: "BOTH" for
use only in the first cell that packets are written
into. I also made a copy constructor for Packet so
that two copies are made; one for left and one for
right. I also changed the bounds of iteration
in Ether.draw().